Completions: move UserSearchEngines into utils
The Vomnibar will use it
