+++ lib/utils.js
@@ -376,6 +376,15 @@ const Utils = {
       return true; // Indicate that we will be calling sendResponse, asynchronously.
     });
   },
+
+  // Remove comments and leading/trailing whitespace from a list of lines, and merge lines where the
+  // last character on the preceding line is "\".
+  parseLines(text) {
+    return text.replace(/\\\n/g, "")
+      .split("\n")
+      .map((line) => line.trim())
+      .filter((line) => (line.length > 0) && !(Array.from('#"').includes(line[0])));
+  },
 };
 
 // This creates a new function out of an existing function, where the new function takes fewer
@@ -534,4 +543,51 @@ class JobRunner {
   }
 }
 
-Object.assign(globalThis, { Utils, SimpleCache, AsyncDataFetcher, JobRunner, EventDispatcher });
+// A struct representing a search engine entry in the "searchEngine" setting.
+class UserSearchEngine {
+  keyword;
+  url;
+  description;
+  constructor(o) {
+    Object.seal(this);
+    if (o) Object.assign(this, o);
+  }
+}
+
+// Parses a user's search engine configuration from Settings, and stores the parsed results.
+// TODO(philc): Should this be responsible for updating itself when Settings changes, rather than
+// the callers doing so? Or, remove this class and re-parse the configuration every keystroke in
+// Vomnibar, so we don't introduce another layer of caching in the code.
+const UserSearchEngines = {
+  keywordToEngine: {},
+
+  parseConfig(configText) {
+    const results = {};
+    for (const line of Utils.parseLines(configText)) {
+      const tokens = line.split(/\s+/);
+      if (tokens.length < 2) continue;
+      if (!tokens[0].includes(":")) continue;
+      const keyword = tokens[0].split(":")[0];
+      const url = tokens[1];
+      const description = tokens.length > 2 ? tokens.slice(2).join(" ") : `search (${keyword})`;
+      if (Utils.hasFullUrlPrefix(url) || Utils.hasJavascriptPrefix(url)) {
+        results[keyword] = new UserSearchEngine({ keyword, url, description });
+      }
+    }
+    return results;
+  },
+
+  set(searchEnginesConfigText) {
+    this.keywordToEngine = this.parseConfig(searchEnginesConfigText);
+  },
+};
+
+Object.assign(globalThis, {
+  Utils,
+  SimpleCache,
+  AsyncDataFetcher,
+  JobRunner,
+  EventDispatcher,
+  UserSearchEngine,
+  UserSearchEngines,
+});
